+++ src/main/java/com/depromeet/domain/reaction/api/ReactionController.java
@@ -1,20 +1,15 @@
 package com.depromeet.domain.reaction.api;
 
 import com.depromeet.domain.reaction.application.ReactionService;
-import com.depromeet.domain.reaction.dto.ReactionCreateRequest;
-import com.depromeet.domain.reaction.dto.ReactionCreateResponse;
+import com.depromeet.domain.reaction.dto.request.ReactionCreateRequest;
+import com.depromeet.domain.reaction.dto.response.ReactionCreateResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @Tag(name = "6. [리액션]", description = "리액션 관련 API")
 @RestController
