+++ src/test/java/com/depromeet/domain/reaction/application/ReactionServiceTest.java
@@ -14,8 +14,8 @@ import com.depromeet.domain.missionRecord.dao.MissionRecordRepository;
 import com.depromeet.domain.missionRecord.domain.MissionRecord;
 import com.depromeet.domain.reaction.dao.ReactionRepository;
 import com.depromeet.domain.reaction.domain.EmojiType;
-import com.depromeet.domain.reaction.dto.ReactionCreateRequest;
-import com.depromeet.domain.reaction.dto.ReactionCreateResponse;
+import com.depromeet.domain.reaction.dto.request.ReactionCreateRequest;
+import com.depromeet.domain.reaction.dto.response.ReactionCreateResponse;
 import com.depromeet.global.error.exception.CustomException;
 import com.depromeet.global.error.exception.ErrorCode;
 import com.depromeet.global.security.PrincipalDetails;
